clean up code

+++ src/main/java/com/sejapoe/techcolonies/core/helper/EntityHelper.java
@@ -5,6 +5,6 @@ import net.minecraft.world.entity.PathfinderMob;
 
 public class EntityHelper {
   public static void moveMobToBlock(PathfinderMob mob, BlockPos pos) {
-    mob.getNavigation().moveTo((double)((float)pos.getX()), pos.getY(), (double)((float)pos.getZ()), 1.0f);
+    mob.getNavigation().moveTo((float)pos.getX(), pos.getY(), (float)pos.getZ(), 1.0f);
   }
 }
